+++ include/HVAC/heater.h
@@ -22,7 +22,7 @@
 #ifndef HEATER_H
 #define HEATER_H
 
-#include "can/CANSPI.h"
+#include "CANSPI.h"
 #include "digio.h"
 #include "my_fp.h"
 #include "my_math.h"
