+++ src/stm32_vcu.cpp
@@ -26,9 +26,9 @@
 #include "vehicles/BMW_E65.h"
 #include "chargers/C5Charger.h"
 #include "dcdc/C5DCDC.h"
-#include "can/CANSPI.h"
+#include "CANSPI.h"
 #include "chargers/CPC.h"
-#include "can/Can_OBD2.h"
+#include "Can_OBD2.h"
 #include "inverters/Can_OI.h"
 #include "vehicles/Can_VAG.h"
 #include "shifters/E65_Lever.h"
