+++ stm32-vcu.cscope_file_list
@@ -36,7 +36,7 @@
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/include/printf.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/utils.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/stm32_vcu.cpp"
-"/home/johu/Projekte/openinverter/Stm32-vcu/src/can/CANSPI.cpp"
+"/home/johu/Projekte/openinverter/Stm32-vcu/src/CANSPI.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/utils.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/chargers/extCharger.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/vehicles/vehicle.h"
@@ -56,14 +56,14 @@
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/src/stm32_can.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/vehicles/BMW_E65.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/vehicles/BMW_E39.h"
-"/home/johu/Projekte/openinverter/Stm32-vcu/src/can/Can_OBD2.cpp"
+"/home/johu/Projekte/openinverter/Stm32-vcu/src/Can_OBD2.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/vehicles/Can_VAG.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/hwinit.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/include/canhardware.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/chargers/OutlanderHeartBeat.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/include/errormessage.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/include/digio.h"
-"/home/johu/Projekte/openinverter/Stm32-vcu/include/can/MCP2515.h"
+"/home/johu/Projekte/openinverter/Stm32-vcu/include/MCP2515.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/throttle.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/vehicles/JLR_G1.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/Makefile"
@@ -100,13 +100,13 @@
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/chargers/amperacharger.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/vehicles/vag_sbox.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/chargers/teslaCharger.cpp"
-"/home/johu/Projekte/openinverter/Stm32-vcu/include/can/CANSPI.h"
+"/home/johu/Projekte/openinverter/Stm32-vcu/include/CANSPI.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/hwinit.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/bms/kangoobms.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/inverters/i3LIM.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/shifters/E65_Lever.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/include/terminal.h"
-"/home/johu/Projekte/openinverter/Stm32-vcu/src/can/MCP2515.cpp"
+"/home/johu/Projekte/openinverter/Stm32-vcu/src/MCP2515.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/src/my_fp.c"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/param_prj.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/vehicles/BMW_E31.h"
@@ -118,7 +118,7 @@
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/NoInverter.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/libopeninv/include/canmap.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/digio_prj.h"
-"/home/johu/Projekte/openinverter/Stm32-vcu/include/can/Can_OBD2.h"
+"/home/johu/Projekte/openinverter/Stm32-vcu/include/Can_OBD2.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/vehicles/JLR_G2.cpp"
 "/home/johu/Projekte/openinverter/Stm32-vcu/include/shifters/F30_Lever.h"
 "/home/johu/Projekte/openinverter/Stm32-vcu/src/vehicles/BMW_E31.cpp"
